[refactor] infrastructure.repository 에 있는 common layer 패키지 레벨 한단계 위로 이동
