+++ src/main/java/junseok/snr/wallet/application/service/TransactionService.java
@@ -1,7 +1,7 @@
 package junseok.snr.wallet.application.service;
 
 import jakarta.persistence.criteria.Predicate;
-import junseok.snr.wallet.infrastructure.repository.common.Web3jUtils;
+import junseok.snr.wallet.infrastructure.common.Web3jUtils;
 import junseok.snr.wallet.api.dto.WithdrawDto;
 import junseok.snr.wallet.domain.model.Transaction;
 import junseok.snr.wallet.domain.model.TransactionType;
