+++ src/main/java/junseok/snr/wallet/infrastructure/common/Web3jUtils.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.infrastructure.repository.common;
+package junseok.snr.wallet.infrastructure.common;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
